+++ example/lib/samples/array_sample.dart
@@ -16,6 +16,9 @@ class _ArraySampleState extends State<ArraySample> {
   FormArray<bool> get selectedContacts =>
       form.control('selectedContacts') as FormArray<bool>;
 
+  FormControl<bool> selectedContactsItem(int i) =>
+      form.control('selectedContacts.$i') as FormControl<bool>;
+
   @override
   void initState() {
     selectedContacts.addAll(
@@ -30,6 +33,8 @@ class _ArraySampleState extends State<ArraySample> {
       children: [
         Expanded(
           child: ReactiveCheckboxListTile(
+            // the use of a Key here, is extremely important
+            key: ValueKey(contact),
             formControlName: contacts.indexOf(contact).toString(),
             title: Text(contact),
           ),
@@ -76,6 +81,34 @@ class _ArraySampleState extends State<ArraySample> {
                           }
                         : null,
                     child: const Text('Send Email'),
+                  );
+                },
+              ),
+              ReactiveFormConsumer(
+                builder: (context, form, child) {
+                  return ElevatedButton(
+                    onPressed: () => selectedContacts.reset(),
+                    child: const Text('reset'),
+                  );
+                },
+              ),
+              ReactiveFormConsumer(
+                builder: (context, form, child) {
+                  return ElevatedButton(
+                    onPressed: () => selectedContacts.updateValue(
+                        selectedContacts.value
+                            ?.map((e) => !(e ?? false))
+                            .toList()),
+                    child: const Text('toggle all'),
+                  );
+                },
+              ),
+              ReactiveFormConsumer(
+                builder: (context, form, child) {
+                  return ElevatedButton(
+                    onPressed: () => selectedContactsItem(0)
+                        .updateValue(!(selectedContactsItem(0).value ?? false)),
+                    child: const Text('toggle first item'),
                   );
                 },
               ),