+++ lib/src/widgets/reactive_date_picker.dart
@@ -77,6 +77,10 @@ class ReactiveDatePicker<T> extends ReactiveFormField<T, DateTime> {
     String? fieldHintText,
     String? fieldLabelText,
     Widget? child,
+    DateTime? currentDate,
+    DateTime? initialDate,
+    TextInputType? keyboardType,
+    Offset? anchorPoint,
   }) : super(
           key: key,
           formControl: formControl,
@@ -88,7 +92,8 @@ class ReactiveDatePicker<T> extends ReactiveFormField<T, DateTime> {
                 field,
                 (field) => showDatePicker(
                   context: field.context,
-                  initialDate: _getInitialDate(field.value, lastDate),
+                  initialDate: _getInitialDate(firstDate, lastDate,
+                      initialDate ?? field.value ?? DateTime.now()),
                   firstDate: firstDate,
                   lastDate: lastDate,
                   initialEntryMode: initialEntryMode,
@@ -106,6 +111,9 @@ class ReactiveDatePicker<T> extends ReactiveFormField<T, DateTime> {
                   errorInvalidText: errorInvalidText,
                   fieldHintText: fieldHintText,
                   fieldLabelText: fieldLabelText,
+                  currentDate: currentDate,
+                  keyboardType: keyboardType,
+                  anchorPoint: anchorPoint,
                 ).then((value) {
                   if (value != null) {
                     field.didChange(value);
@@ -117,13 +125,17 @@ class ReactiveDatePicker<T> extends ReactiveFormField<T, DateTime> {
           },
         );
 
-  static DateTime _getInitialDate(DateTime? fieldValue, DateTime lastDate) {
-    if (fieldValue != null) {
-      return fieldValue;
+  static DateTime _getInitialDate(
+      DateTime firstDate, DateTime lastDate, DateTime defaultInitDate) {
+    var initialDate = defaultInitDate;
+
+    if (initialDate.isBefore(firstDate)) {
+      initialDate = firstDate;
+    } else if (initialDate.isAfter(lastDate)) {
+      initialDate = lastDate;
     }
 
-    final now = DateTime.now();
-    return now.compareTo(lastDate) > 0 ? lastDate : now;
+    return initialDate;
   }
 
   @override