+++ lib/src/widgets/reactive_radio.dart
@@ -12,7 +12,7 @@ import 'package:reactive_forms/reactive_forms.dart';
 /// to a [FormControl].
 ///
 /// For documentation about the various parameters, see the [Radio] class
-/// and [new Radio], the constructor.
+/// and [Radio], the constructor.
 class ReactiveRadio<T> extends ReactiveFormField<T, T> {
   /// Creates a [ReactiveRadio] that contains a [Radio].
   ///
@@ -27,7 +27,7 @@ class ReactiveRadio<T> extends ReactiveFormField<T, T> {
   /// The [value] arguments is required. See [Radio] constructor.
   ///
   /// For documentation about the various parameters, see the [Radio] class
-  /// and [new Radio], the constructor.
+  /// and [Radio], the constructor.
   ReactiveRadio({
     Key? key,
     String? formControlName,
@@ -44,11 +44,16 @@ class ReactiveRadio<T> extends ReactiveFormField<T, T> {
     double? splashRadius,
     bool autofocus = false,
     bool toggleable = false,
+    FocusNode? focusNode,
   }) : super(
           key: key,
           formControl: formControl,
           formControlName: formControlName,
-          builder: (ReactiveFormFieldState<T, T> field) {
+          builder: (field) {
+            final state = field as _ReactiveRadioState<T, T>;
+
+            state._setFocusNode(focusNode);
+
             return Radio<T>(
               value: value,
               groupValue: field.value,
@@ -64,11 +69,48 @@ class ReactiveRadio<T> extends ReactiveFormField<T, T> {
               visualDensity: visualDensity,
               autofocus: autofocus,
               toggleable: toggleable,
-              // focusNode: focusNode - requires more time
+              focusNode: state.focusNode,
             );
           },
         );
 
   @override
-  ReactiveFormFieldState<T, T> createState() => ReactiveFormFieldState<T, T>();
+  ReactiveFormFieldState<T, T> createState() => _ReactiveRadioState<T, T>();
+}
+
+class _ReactiveRadioState<T, V> extends ReactiveFormFieldState<T, V> {
+  FocusNode? _focusNode;
+  late FocusController _focusController;
+
+  FocusNode get focusNode => _focusNode ?? _focusController.focusNode;
+
+  @override
+  void subscribeControl() {
+    _registerFocusController(FocusController());
+    super.subscribeControl();
+  }
+
+  @override
+  void unsubscribeControl() {
+    _unregisterFocusController();
+    super.unsubscribeControl();
+  }
+
+  void _registerFocusController(FocusController focusController) {
+    _focusController = focusController;
+    control.registerFocusController(focusController);
+  }
+
+  void _unregisterFocusController() {
+    control.unregisterFocusController(_focusController);
+    _focusController.dispose();
+  }
+
+  void _setFocusNode(FocusNode? focusNode) {
+    if (_focusNode != focusNode) {
+      _focusNode = focusNode;
+      _unregisterFocusController();
+      _registerFocusController(FocusController(focusNode: _focusNode));
+    }
+  }
 }