+++ lib/src/widgets/reactive_radio_list_tile.dart
@@ -12,7 +12,7 @@ import 'package:reactive_forms/reactive_forms.dart';
 /// to a [FormControl].
 ///
 /// For documentation about the various parameters, see the [RadioListTile]
-/// class and [new RadioListTile], the constructor.
+/// class and [RadioListTile], the constructor.
 class ReactiveRadioListTile<T> extends ReactiveFormField<T, T> {
   /// Create an instance of a [ReactiveRadioListTile].
   ///
@@ -44,11 +44,18 @@ class ReactiveRadioListTile<T> extends ReactiveFormField<T, T> {
     ShapeBorder? shape,
     bool autofocus = false,
     bool selected = false,
+    VisualDensity? visualDensity,
+    FocusNode? focusNode,
+    bool? enableFeedback,
   }) : super(
           key: key,
           formControl: formControl,
           formControlName: formControlName,
-          builder: (ReactiveFormFieldState<T, T> field) {
+          builder: (field) {
+            final state = field as _ReactiveRadioListTileState<T, T>;
+
+            state._setFocusNode(focusNode);
+
             return RadioListTile<T>(
               value: value,
               groupValue: field.value,
@@ -67,10 +74,51 @@ class ReactiveRadioListTile<T> extends ReactiveFormField<T, T> {
               shape: shape,
               selected: selected,
               autofocus: autofocus,
+              visualDensity: visualDensity,
+              focusNode: state.focusNode,
+              enableFeedback: enableFeedback,
             );
           },
         );
 
   @override
-  ReactiveFormFieldState<T, T> createState() => ReactiveFormFieldState<T, T>();
+  ReactiveFormFieldState<T, T> createState() =>
+      _ReactiveRadioListTileState<T, T>();
+}
+
+class _ReactiveRadioListTileState<T, V> extends ReactiveFormFieldState<T, V> {
+  FocusNode? _focusNode;
+  late FocusController _focusController;
+
+  FocusNode get focusNode => _focusNode ?? _focusController.focusNode;
+
+  @override
+  void subscribeControl() {
+    _registerFocusController(FocusController());
+    super.subscribeControl();
+  }
+
+  @override
+  void unsubscribeControl() {
+    _unregisterFocusController();
+    super.unsubscribeControl();
+  }
+
+  void _registerFocusController(FocusController focusController) {
+    _focusController = focusController;
+    control.registerFocusController(focusController);
+  }
+
+  void _unregisterFocusController() {
+    control.unregisterFocusController(_focusController);
+    _focusController.dispose();
+  }
+
+  void _setFocusNode(FocusNode? focusNode) {
+    if (_focusNode != focusNode) {
+      _focusNode = focusNode;
+      _unregisterFocusController();
+      _registerFocusController(FocusController(focusNode: _focusNode));
+    }
+  }
 }