+++ test/src/widgets/reactive_date_picker_testing_widget.dart
@@ -4,23 +4,32 @@ import 'package:reactive_forms/reactive_forms.dart';
 class ReactiveDatePickerTestingWidget<T> extends StatelessWidget {
   final FormGroup form;
   final DateTime? lastDate;
+  final DateTime? firstDate;
+  final DateTime? initialDate;
 
   const ReactiveDatePickerTestingWidget({
     Key? key,
     required this.form,
     this.lastDate,
+    this.firstDate,
+    this.initialDate,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    final now = DateTime.now();
+    final firstDate = this.firstDate ?? now.subtract(const Duration(days: 365));
+    final lastDate = this.lastDate ?? now.add(const Duration(days: 365));
+
     return MaterialApp(
       home: Material(
         child: ReactiveForm(
           formGroup: form,
           child: ReactiveDatePicker<T>(
             formControlName: 'birthday',
-            firstDate: DateTime(1985),
-            lastDate: lastDate ?? DateTime(2050),
+            initialDate: initialDate,
+            firstDate: firstDate,
+            lastDate: lastDate,
             builder: (BuildContext context,
                 ReactiveDatePickerDelegate<T> picker, Widget? child) {
               return TextButton(